# docs/reference/openapi.yaml
openapi: 3.1.0
info:
  title: thelake API
  version: 0.1.0
  description: |
    Authenticated HTTP API exposed by thelake. Interactive UI: `GET /swagger`
    (this document is also served as `GET /openapi.json`).

    This API accepts telemetry payloads on standard OTLP HTTP paths:
    - `POST /v1/traces`
    - `POST /v1/logs`

    It also documents the tenant-scoped schema promotion endpoint:
    - `POST /v1/promotions/apply`

    Evaluations are written through:
    - `POST /v1/scores`
    - `GET|POST /v1/score-configs`

    Query surfaces:
    - `POST /v1/traces/search`
    - `POST /v1/traces/details`
    - `POST /v1/spans/search`
    - `GET /v1/spans/{span_id}`
    - `GET /v1/traces/{trace_id}`
    - `POST /v1/sessions/search`
    - `POST /v1/sessions/details`
    - `POST /v1/sessions/summary/rebuild`
    - `GET /v1/sessions/{session_id}`
    - `GET /v1/sessions/{session_id}/recording`

    Control plane:
    - `POST /v1/workspaces` (admin bearer)
    - `GET /v1/meta`
    - `POST /v1/promotions/apply`
    - `GET /v1/data/ducklake-connection`

    Also inventoried here (summary-level): health/ready, `POST /v1/query/sql`,
    and `/v1/fields`. Loki/Tempo Grafana-compat query routes are served by
    the process but are not part of this product OpenAPI document.

    For JSON payloads, requests are expected to follow the OpenTelemetry collector export
    request shapes (`ExportTraceServiceRequest` and `ExportLogsServiceRequest`).

...

...

...

...
    `docs/how-to/promotion.md`.
...
servers:
...
security:
  - bearerAuth: []
paths:
  /slack/events:
    post:
      tags: [integrations]
      summary: Receive Slack Events API callbacks
      description: |
        Slack verifies this endpoint using the configured signing secret. It
        accepts URL verification challenges and authorized evaluator commands
        from Slack event callbacks. This endpoint does not use bearer auth.
      operationId: receiveSlackEvent
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [type]
              additionalProperties: true
              description: Slack Events API envelope, signed over the raw body.
      responses:
        "200":
          description: Challenge response or event acknowledged
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        "400":
          description: Malformed Slack event envelope
        "401":
          description: Missing, stale, or invalid Slack request signature
        "403":
          description: Slack team or author is not authorized
        "503":
          description: Slack integration is unavailable; Slack may retry the event
...
  /v1/traces:
    post:
      tags:
        - ingestion
      summary: Ingest OTLP traces
      operationId: ingestTraces
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExportTraceServiceRequest"
          application/x-protobuf:
            schema:
              type: string
              format: binary
              description: Protobuf-encoded `ExportTraceServiceRequest`
      responses:
        "200":
          description: Request accepted and processed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IngestResponse"
        "400":
          description: Invalid JSON/protobuf payload
          content:
            text/plain:
              schema:
                $ref: "#/components/schemas/ErrorMessage"
              examples:
                invalidJson:
                  value: Invalid JSON
                invalidProtobuf:
                  value: Protobuf decode failed
        "401":
          description: Missing or invalid bearer token
        "403":
          description: Bearer token could not be resolved to a tenant
        "413":
          description: Request exceeds `server.max_body_size`
        "503":
          $ref: "#/components/responses/IngestUnavailable"
  /v1/logs:
    post:
      tags:
        - ingestion
      summary: Ingest OTLP logs
      operationId: ingestLogs
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExportLogsServiceRequest"
          application/x-protobuf:
            schema:
              type: string
              format: binary
              description: Protobuf-encoded `ExportLogsServiceRequest`
      responses:
        "200":
          description: Request accepted and processed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IngestResponse"
        "400":
          description: Invalid JSON/protobuf payload
          content:
            text/plain:
              schema:
                $ref: "#/components/schemas/ErrorMessage"
              examples:
                invalidJson:
                  value: "Invalid JSON: expected value at line 1 column 1"
                invalidProtobuf:
                  value: Protobuf decode failed
        "401":
          description: Missing or invalid bearer token
        "403":
          description: Bearer token could not be resolved to a tenant
        "413":
          description: Request exceeds `server.max_body_size`
        "503":
          $ref: "#/components/responses/IngestUnavailable"
  /v1/scores:
    post:
      tags:
        - llm
      summary: Create an immutable LLM evaluation score
      description: |
        Creates a score in the authenticated tenant's DuckLake scope. Tenant
        identity is resolved from the bearer token and is never accepted in
        the request body.

        At least one target (`trace_id`, `span_id`, or `session_id`) is
        required. Exactly one value field must be present and must match
        `data_type`:

        - `numeric` uses `numeric_value`
        - `categorical` and `text` use `string_value`
        - `boolean` uses `boolean_value`

        `score_id` is the tenant-local idempotency key. Repeating a previously
        accepted ID does not append another score row.
      operationId: createScore
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateScoreRequest"
            examples:
              correctness:
                value:
                  score_id: score-01JZ8YH3KQBZ8V2CA9ZB1W7Q9A
                  timestamp: "2026-07-18T23:22:00Z"
                  trace_id: 4bf92f3577b34da6a3ce929d0e0e4736
                  span_id: 00f067aa0ba902b7
                  name: correctness
                  data_type: numeric
                  numeric_value: 0.92
                  source: evaluator
                  comment: Answer is supported by the retrieved context.
                  metadata:
                    evaluator: groundedness-v2
      responses:
        "201":
          description: Score created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Score"
        "200":
          description: Idempotent retry; the score ID already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Score"
        "400":
          description: Missing target, missing value, or value does not match `data_type`
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "401":
          description: Missing or invalid bearer token
        "403":
          description: Bearer token could not be resolved to a tenant
        "422":
          description: Malformed JSON or field type mismatch
        "503":
          description: Tenant runtime, idempotency lookup, or DuckLake write unavailable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /v1/policies:
    get:
      tags: [llm]
      summary: List workspace and agent Markdown policy versions
      description: Returns workspace-scoped immutable Markdown policy versions.
      operationId: listPolicies
      responses:
        "200":
          description: Policy documents
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PolicyDocument"
        "401":
          description: Missing or invalid bearer token
        "403":
          description: Bearer token could not be resolved to a tenant
    post:
      tags: [llm]
      summary: Create an immutable Markdown policy version
      description: Stores workspace-wide or exact-agent policy content in the authenticated workspace.
      operationId: createPolicy
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePolicyRequest"
      responses:
        "201":
          description: Policy version created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PolicyDocument"
        "200":
          description: Idempotent retry of the same policy version
        "400":
          description: Invalid policy or conflicting immutable version
        "401":
          description: Missing or invalid bearer token
        "403":
          description: Bearer token could not be resolved to a tenant
  /v1/evaluators:
    get:
      tags: [llm]
      summary: List no-code behavior evaluators
      description: Returns versioned evaluator criteria stored in the authenticated workspace.
      operationId: listEvaluators
      responses:
        "200":
          description: Evaluator definitions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/EvaluatorDefinition"
        "401":
          description: Missing or invalid bearer token
        "403":
          description: Bearer token could not be resolved to a tenant
    post:
      tags: [llm]
      summary: Create a draft no-code behavior evaluator version
      description: |
        Stores a natural-language criterion and optional ordered tool
        requirement for one named agent in the authenticated workspace. Drafts
        do not trigger evaluations. Explicit activation opts this evaluator
        into sending matching trace content to the configured judge provider.
      operationId: createEvaluator
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateEvaluatorRequest"
            examples:
              eligibilityBeforeRefund:
                value:
                  evaluator_id: refund-eligibility
                  version: 1
                  target_agent_name: support-agent
                  name: Refund eligibility before refund
                  criteria: Verify the agent checks refund eligibility and explains the result before issuing a refund.
                  threshold: 0.7
                  uncertainty_margin: 0.1
                  required_tool_order:
                    - before: check_eligibility
                      action: issue_refund
                      require_result_before_action: true
      responses:
        "201":
          description: Evaluator version created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EvaluatorDefinition"
        "200":
          description: Idempotent retry of the same evaluator version
        "400":
          description: Invalid evaluator definition or conflicting config ID
        "401":
          description: Missing or invalid bearer token
        "403":
          description: Bearer token could not be resolved to a tenant
  /v1/evaluators/{evaluator_id}/versions/{version}/activate:
    post:
      tags: [llm]
      summary: Activate an evaluator version for one agent
      description: |
        Activates this workspace-scoped evaluator for new matching traces. When
        the DeepEval runner is configured, captured prompt, completion, and
        tool data for the selected agent is sent to the configured Gemini
        provider. The runner applies a narrow credential redaction filter;
        this is not general PII scrubbing.
      operationId: activateEvaluator
      parameters:
        - name: evaluator_id
          in: path
          required: true
          schema: { type: string }
        - name: version
          in: path
          required: true
          schema: { type: integer, minimum: 1 }
      responses:
        "200":
          description: Evaluator activated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EvaluatorDefinition"
        "400":
          description: Evaluator version not found
        "503":
          description: Online evaluation runner is not configured
        "401":
          description: Missing or invalid bearer token
        "403":
          description: Bearer token could not be resolved to a tenant
  /v1/evaluators/{evaluator_id}/versions/{version}/deactivate:
    post:
      tags: [llm]
      summary: Stop evaluating new traces with this evaluator
      operationId: deactivateEvaluator
      parameters:
        - name: evaluator_id
          in: path
          required: true
          schema: { type: string }
        - name: version
          in: path
          required: true
          schema: { type: integer, minimum: 1 }
      responses:
        "200":
          description: Evaluator deactivated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EvaluatorDefinition"
        "400":
          description: Evaluator version not found
        "401":
          description: Missing or invalid bearer token
        "403":
          description: Bearer token could not be resolved to a tenant
  /v1/score-configs:
    get:
      tags:
        - llm
      summary: List score configs (seeds defaults when empty)
      description: |
        Returns append-only score schemas for the authenticated tenant. When the
        tenant has no configs yet, the server seeds built-in defaults
        (`correctness`, `quality`, `expected_output`) and returns them.
      operationId: listScoreConfigs
      responses:
        "200":
          description: Score config list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScoreConfigListResponse"
        "401":
          description: Missing or invalid bearer token
        "403":
          description: Bearer token could not be resolved to a tenant
        "503":
          description: Tenant runtime or DuckLake unavailable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
    post:
      tags:
        - llm
      summary: Create an append-only score config
      description: |
        Creates a score config in the authenticated tenant's DuckLake scope.
        `config_id` is the tenant-local idempotency key.
      operationId: createScoreConfig
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateScoreConfigRequest"
            examples:
              correctness:
                value:
                  config_id: cfg-correctness
                  timestamp: "2026-07-18T23:22:00Z"
                  name: correctness
                  data_type: boolean
                  description: Whether the turn was correct
                  author_id: system
      responses:
        "201":
          description: Score config created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScoreConfig"
        "200":
          description: Existing config returned after an idempotent retry
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScoreConfig"
        "400":
          description: Invalid score config
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "401":
          description: Missing or invalid bearer token
        "403":
          description: Bearer token could not be resolved to a tenant
        "503":
          description: Tenant runtime or DuckLake write unavailable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /v1/spans/search:
    post:
      tags:
        - llm
      summary: Search lightweight LLM span projections
      description: |
        Requires `from` and `to`. Returns lightweight span rows without
        full attributes/events. LLM fields are derived from span attributes
        (`gen_ai.*`, `sp.*`) and do not require promotion to be applied.
        Cursor ordering is `(timestamp DESC, span_id DESC)`.
      operationId: searchSpans
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SpanSearchRequest"
      responses:
        "200":
          description: Bounded span page
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SpanSearchResponse"
        "400":
          description: Missing time range, invalid filter, or malformed cursor
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "401":
          description: Missing or invalid bearer token
        "403":
          description: Bearer token could not be resolved to a tenant
        "503":
          description: Tenant runtime or query unavailable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /v1/spans/{span_id}:
    get:
      tags:
        - llm
      summary: Get one span with full payload and attached scores
      operationId: getSpan
      parameters:
        - name: span_id
          in: path
          required: true
          schema:
            type: string
        - name: from
          in: query
          required: false
          schema:
            type: string
            format: date-time
        - name: to
          in: query
          required: false
          schema:
            type: string
            format: date-time
      responses:
        "200":
          description: Span detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SpanDetail"
        "404":
          description: Span not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "401":
          description: Missing or invalid bearer token
        "503":
          description: Tenant runtime or query unavailable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /v1/traces/{trace_id}:
    get:
      tags:
        - llm
      summary: Get a trace, its spans, and scores
      operationId: getTrace
      parameters:
        - name: trace_id
          in: path
          required: true
          schema:
            type: string
        - name: from
          in: query
          required: false
          schema:
            type: string
            format: date-time
        - name: to
          in: query
          required: false
          schema:
            type: string
            format: date-time
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 200
            default: 100
        - name: cursor
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: Trace detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TraceDetail"
        "400":
          description: Invalid limit or cursor
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "404":
          description: Trace not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "401":
          description: Missing or invalid bearer token
        "503":
          description: Tenant runtime or query unavailable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /v1/sessions/search:
    post:
      tags:
        - llm
      summary: Search sessions via session_summary
      description: |
        Always backed by Postgres `session_summary` (empty table → empty page).
        Requires `from` and `to`. Cursor paging is only defined for
        `order_by=start_time`; other orderings return a single ranked page and
        set `cursor_supported=false`.
      operationId: searchSessions
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SessionSearchRequest"
      responses:
        "200":
          description: Bounded session page
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SessionSearchResponse"
        "400":
          description: Missing time range, invalid filter, or malformed cursor
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "401":
          description: Missing or invalid bearer token
        "403":
          description: Bearer token could not be resolved to a tenant
        "503":
          description: session_summary or tenant runtime unavailable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /v1/sessions/summary/rebuild:
    post:
      tags:
        - llm
      summary: Rebuild session_summary for an explicit time window
      description: |
        Ops endpoint. Rejects inverted or oversized windows, acquires a
        per-tenant rebuild lease, then runs the shared lake aggregate → UPSERT
        path.
      operationId: rebuildSessionSummary
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SessionSummaryRebuildRequest"
      responses:
        "200":
          description: Rebuild completed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SessionSummaryRebuildResponse"
        "400":
          description: Invalid or oversized rebuild window
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "401":
          description: Missing or invalid bearer token
        "503":
          description: Rebuild lease, lake aggregate, or UPSERT unavailable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /v1/sessions/{session_id}:
    get:
      tags:
        - llm
      summary: Get full session detail using the session summary time window
      description: |
        Does not accept query `from`/`to`. The lake window is taken from the
        tenant's `session_summary` row for `session_id`. Returns all session spans,
        trace and token totals, and scores in one response.
      operationId: getSession
      parameters:
        - name: session_id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Complete session detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SessionDetail"
        "400":
          description: Invalid session ID or query
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "404":
          description: No session_summary row or no lake activity in the summary window
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "401":
          description: Missing or invalid bearer token
        "503":
          description: session_summary unavailable or tenant runtime unavailable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /v1/sessions/{session_id}/recording:
    get:
      tags:
        - llm
      summary: Get web session recording batches for a session
      operationId: getSessionRecording
      parameters:
        - name: session_id
          in: path
          required: true
          schema:
            type: string
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 200
            default: 50
      responses:
        "200":
          description: Ordered recording batches and flattened rrweb events
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SessionRecording"
        "400":
          description: Invalid limit
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "404":
          description: No session_summary row for this session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "401":
          description: Missing or invalid bearer token
        "503":
          description: session_summary unavailable or tenant runtime unavailable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
...
  /v1/fields/{field}/values:
    get:
      tags: [query]
      summary: List values for a field
      operationId: listFieldValues
      parameters:
        - name: field
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Field values
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  responses:
    IngestUnavailable:
      description: DuckLake commit failed after backend retries; exporter should retry
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/IngestResponse"
          example:
            success: false
            ingested_count: 0
            message: "Ingestion failed: DuckLake write failed"
  schemas:
...
    ErrorMessage:
      type: string
      description: Plain-text error body returned for malformed requests.
    ExportTraceServiceRequest:
      type: object
      description: |
        OTLP trace export request (JSON mapping).
        See OpenTelemetry protocol spec for the canonical schema.
      additionalProperties: true
...